// lib/graphql/types/v2/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Values that can be edited in Account's settings */
  AccountSettingsKey: any;
  DateString: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: any;
  IsoDateString: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
  /** A string that cannot be passed as an empty value */
  NonEmptyString: any;
  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
  URL: any;
};
// ...
export type Amount = {
  __typename?: 'Amount';
  currency?: Maybe<Currency>;
  /** If the amount was generated from a currency conversion, this field contains details about the conversion */
  exchangeRate?: Maybe<CurrencyExchangeRate>;
  value?: Maybe<Scalars['Float']>;
  valueInCents?: Maybe<Scalars['Float']>;
};
// ...
export enum Currency {
  /** UAE Dirham */
  AED = 'AED',
  /** Afghani */
  AFN = 'AFN',
  /** Lek */
  ALL = 'ALL',
  /** Armenian Dram */
  AMD = 'AMD',
  /** Netherlands Antillean Guilder */
  ANG = 'ANG',
  /** Kwanza */
  AOA = 'AOA',
  /** Argentine Peso */
  ARS = 'ARS',
  /** Australian Dollar */
  AUD = 'AUD',
  /** Aruban Florin */
  AWG = 'AWG',
  /** Azerbaijanian Manat */
  AZN = 'AZN',
  /** Convertible Mark */
  BAM = 'BAM',
  /** Barbados Dollar */
  BBD = 'BBD',
  /** Taka */
  BDT = 'BDT',
  /** Bulgarian Lev */
  BGN = 'BGN',
  /** Burundi Franc */
  BIF = 'BIF',
  /** Bermudian Dollar */
  BMD = 'BMD',
  /** Brunei Dollar */
  BND = 'BND',
  /** Boliviano */
  BOB = 'BOB',
  /** Brazilian Real */
  BRL = 'BRL',
  /** Bahamian Dollar */
  BSD = 'BSD',
  /** Pula */
  BWP = 'BWP',
  /** Belarussian Ruble */
  BYN = 'BYN',
  /** Belize Dollar */
  BZD = 'BZD',
  /** Canadian Dollar */
  CAD = 'CAD',
  /** Congolese Franc */
  CDF = 'CDF',
  /** Swiss Franc */
  CHF = 'CHF',
  /** Chilean Peso */
  CLP = 'CLP',
  /** Yuan Renminbi */
  CNY = 'CNY',
  /** Colombian Peso */
  COP = 'COP',
  /** Costa Rican Colon */
  CRC = 'CRC',
  /** Cabo Verde Escudo */
  CVE = 'CVE',
  /** Czech Koruna */
  CZK = 'CZK',
  /** Djibouti Franc */
  DJF = 'DJF',
  /** Danish Krone */
  DKK = 'DKK',
  /** Dominican Peso */
  DOP = 'DOP',
  /** Algerian Dinar */
  DZD = 'DZD',
  /** Egyptian Pound */
  EGP = 'EGP',
  /** Ethiopian Birr */
  ETB = 'ETB',
  /** Euro */
  EUR = 'EUR',
  /** Fiji Dollar */
  FJD = 'FJD',
  /** Falkland Islands Pound */
  FKP = 'FKP',
  /** Pound Sterling */
  GBP = 'GBP',
  /** Lari */
  GEL = 'GEL',
  /** Gibraltar Pound */
  GIP = 'GIP',
  /** Dalasi */
  GMD = 'GMD',
  /** Guinea Franc */
  GNF = 'GNF',
  /** Quetzal */
  GTQ = 'GTQ',
  /** Guyana Dollar */
  GYD = 'GYD',
  /** Hong Kong Dollar */
  HKD = 'HKD',
  /** Lempira */
  HNL = 'HNL',
  /** Kuna */
  HRK = 'HRK',
  /** Gourde */
  HTG = 'HTG',
  /** Forint */
  HUF = 'HUF',
  /** Rupiah */
  IDR = 'IDR',
  /** New Israeli Sheqel */
  ILS = 'ILS',
  /** Indian Rupee */
  INR = 'INR',
  /** Iceland Krona */
  ISK = 'ISK',
  /** Jamaican Dollar */
  JMD = 'JMD',
  /** Yen */
  JPY = 'JPY',
  /** Kenyan Shilling */
  KES = 'KES',
  /** Som */
  KGS = 'KGS',
  /** Riel */
  KHR = 'KHR',
  /** Comoro Franc */
  KMF = 'KMF',
  /** Won */
  KRW = 'KRW',
  /** Cayman Islands Dollar */
  KYD = 'KYD',
  /** Tenge */
  KZT = 'KZT',
  /** Kip */
  LAK = 'LAK',
  /** Lebanese Pound */
  LBP = 'LBP',
  /** Sri Lanka Rupee */
  LKR = 'LKR',
  /** Liberian Dollar */
  LRD = 'LRD',
  /** Loti */
  LSL = 'LSL',
  /** Moroccan Dirham */
  MAD = 'MAD',
  /** Moldovan Leu */
  MDL = 'MDL',
  /** Malagasy Ariary */
  MGA = 'MGA',
  /** Denar */
  MKD = 'MKD',
  /** Kyat */
  MMK = 'MMK',
  /** Tugrik */
  MNT = 'MNT',
  /** Pataca */
  MOP = 'MOP',
  /** Mauritius Rupee */
  MUR = 'MUR',
  /** Rufiyaa */
  MVR = 'MVR',
  /** Kwacha */
  MWK = 'MWK',
  /** Mexican Peso */
  MXN = 'MXN',
  /** Malaysian Ringgit */
  MYR = 'MYR',
  /** Mozambique Metical */
  MZN = 'MZN',
  /** Namibia Dollar */
  NAD = 'NAD',
  /** Naira */
  NGN = 'NGN',
  /** Cordoba Oro */
  NIO = 'NIO',
  /** Norwegian Krone */
  NOK = 'NOK',
  /** Nepalese Rupee */
  NPR = 'NPR',
  /** New Zealand Dollar */
  NZD = 'NZD',
  /** Balboa */
  PAB = 'PAB',
  /** Nuevo Sol */
  PEN = 'PEN',
  /** Kina */
  PGK = 'PGK',
  /** Philippine Peso */
  PHP = 'PHP',
  /** Pakistan Rupee */
  PKR = 'PKR',
  /** Zloty */
  PLN = 'PLN',
  /** Guarani */
  PYG = 'PYG',
  /** Qatari Rial */
  QAR = 'QAR',
  /** Romanian Leu */
  RON = 'RON',
  /** Serbian Dinar */
  RSD = 'RSD',
  /** Russian Ruble */
  RUB = 'RUB',
  /** Rwanda Franc */
  RWF = 'RWF',
  /** Saudi Riyal */
  SAR = 'SAR',
  /** Solomon Islands Dollar */
  SBD = 'SBD',
  /** Seychelles Rupee */
  SCR = 'SCR',
  /** Swedish Krona */
  SEK = 'SEK',
  /** Singapore Dollar */
  SGD = 'SGD',
  /** Saint Helena Pound */
  SHP = 'SHP',
  /** Leone */
  SLL = 'SLL',
  /** Somali Shilling */
  SOS = 'SOS',
  /** Surinam Dollar */
  SRD = 'SRD',
  /** Lilangeni */
  SZL = 'SZL',
  /** Baht */
  THB = 'THB',
  /** Somoni */
  TJS = 'TJS',
  /** Pa’anga */
  TOP = 'TOP',
  /** Turkish Lira */
  TRY = 'TRY',
  /** Trinidad and Tobago Dollar */
  TTD = 'TTD',
  /** New Taiwan Dollar */
  TWD = 'TWD',
  /** Tanzanian Shilling */
  TZS = 'TZS',
  /** Hryvnia */
  UAH = 'UAH',
  /** Uganda Shilling */
  UGX = 'UGX',
  /** US Dollar */
  USD = 'USD',
  /** Peso Uruguayo */
  UYU = 'UYU',
  /** Uzbekistan Sum */
  UZS = 'UZS',
  /** Dong */
  VND = 'VND',
  /** Vatu */
  VUV = 'VUV',
  /** Tala */
  WST = 'WST',
  /** CFA Franc BEAC */
  XAF = 'XAF',
  /** East Caribbean Dollar */
  XCD = 'XCD',
  /** CFA Franc BCEAO */
  XOF = 'XOF',
  /** CFP Franc */
  XPF = 'XPF',
  /** Yemeni Rial */
  YER = 'YER',
  /** Rand */
  ZAR = 'ZAR',
  /** Zambian Kwacha */
  ZMW = 'ZMW'
}

/** Fields for a currency fx rate */
export type CurrencyExchangeRate = {
  __typename?: 'CurrencyExchangeRate';
  /** Date of the FX rate */
  date: Scalars['DateTime'];
  fromCurrency: Currency;
  /** Is the FX rate approximate or a fixed value? */
  isApproximate: Scalars['Boolean'];
  /** Where does the FX rate comes from */
  source: CurrencyExchangeRateSourceType;
  toCurrency: Currency;
  /** Exchange rate value as a scalar (e.g 1.15 or 0.86) */
  value: Scalars['Float'];
};

/** Where does the FX rate come from */
export enum CurrencyExchangeRateSourceType {
  /** Open Collective internal system, relying on caching and 3rd party APIs */
  OPENCOLLECTIVE = 'OPENCOLLECTIVE',
  /** PayPal API */
  PAYPAL = 'PAYPAL',
  /** Wise API */
  WISE = 'WISE'
}